modify: Condense cookbook api methods
